move IteratorAggregateStub into tests\Fixtures
